+++ src/redturtle/volto/upgrades.py
@@ -508,3 +508,17 @@ def to_4305(context):
     installer.uninstall_product(product_id="collective.volto.cookieconsent")
     if not installer.is_product_installed("collective.volto.gdprcookie"):
         installer.install_product(product_id="collective.volto.gdprcookie")
+
+
+def to_4306(context):
+    portal = api.portal.get()
+    installer = get_installer(portal, portal.REQUEST)
+    if not installer.is_product_installed("collective.volto.sitesettings"):
+        installer.install_product(product_id="collective.volto.sitesettings")
+
+
+def to_4307(context):
+    context.runImportStepFromProfile(
+        "profile-redturtle.volto:profile_to_4307", "plone.app.registry", False
+    )
+    api.portal.set_registry_record("redturtle.volto.rss_image_choice", "image")