+++ CHANGES.rst
@@ -1,16 +1,31 @@
 Changelog
 =========
 
-5.4.10 (unreleased)
--------------------
+5.5.2 (unreleased)
+------------------
+
+- Nothing changed yet.
+
+
+5.5.1 (2024-07-22)
+------------------
+
+- Allow to select which image miniature use in
+  RSS
+  [lucabel]
+
+
+5.5.0 (2024-07-10)
+------------------
 
 - remove newsitem template override, use default dexterity view for newsitem in backend
   [mamico]
 
 - Fixed limit event occurrences to 100.
   [eikichi18]
-
-
+- Add dependency with collective.volto.sitesettings.
+  [cekk]
+  
 5.4.9 (2024-04-22)
 ------------------